ml-service/requirements.txt
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
numpy==2.2.3
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
pymongo==4.11.3
python-dotenv==1.1.0
